# bit-vector: optimize bit-vector-ref and bit-vector-set!