bit-vector-ref and bit-vector-set! are often used in a loop, so it is
worth making them efficient.

The biggest improvement comes from switching from `provide/contract` to
plain `provide` and add manual checks instead.

On a benchmark program, which calculates all prime numbers up to
10_000_000 elements

With `provide/contract`: 954ms
With `provide`: 251ms (this PR)

Furthermore, unsafe operations are employed because they are actually
safe for uses.

With safe operations: 294ms
With unsafe operations: 251ms (this PR)