@rfindler 2022-06-23
Can someone share the specific benchmark that had the big speed up from
removing the contracts, please?

Thanks,
Robby