@sorawee 2022-06-23
@rfindler It's this program:

```
#lang racket/base

(require data/bit-vector
         racket/list)

(define (find-all n)
  (define not-prime-table (make-bit-vector (add1 n)))

  (for/list ([i (in-inclusive-range 2 n)] 
             #:unless (bit-vector-ref not-prime-table i))
    (for ([j (in-inclusive-range (+ i i) n i)])
      (bit-vector-set! not-prime-table j #t))
    i))

(define primes (time (find-all 10000000)))
```