@sorawee 2022-06-23
Here's a simplified program:

```
#lang racket/base

(module sub racket/base
  (require racket/contract/base)
  (define (call a b c)
    (void))

  (define (call/check a b c)
    (unless (boolean? a)
      (error 'bad))
    (unless (boolean? b)
      (error 'bad))
    (unless (boolean? c)
      (error 'bad))
    (void))

  (provide/contract [call (-> boolean? boolean? boolean? any)])
  (provide call/check))

(require 'sub)

(define (loop n)
  (if (zero? n)
      123456789
      (begin
        (call/check #t #t #t)
        (loop (sub1 n)))))

(time (loop 100000000))
```

Results:

- `call/check` with 1 arguments: 239ms
- `call/check` with 2 arguments: 243ms
- `call/check` with 3 arguments: 250ms
- `call/check` with 4 arguments: 269ms
- `call` with 1 arguments: 730ms
- `call` with 2 arguments: 1522ms
- `call` with 3 arguments: 1976ms
- `call` with 4 arguments: 2331ms

The contracted version seems to be adding about half a second for every argument added.