@samth 2022-06-24
I'm not sure exactly how you're running this, but I get basically identical output for the two functions:


```
#lang racket/base

(module sub racket/base
  (require racket/contract/base)
  (define (call a b c)
    (void))

  (define (call/check a b c)
    (unless (boolean? a)
      (error 'bad))
    (unless (boolean? b)
      (error 'bad))
    (unless (boolean? c)
      (error 'bad))
    (void))

  (provide/contract [call (-> boolean? boolean? boolean? any)])
  (provide call/check))

(require 'sub)

(define (loop n)
  (if (zero? n)
      123456789
      (begin
        (call/check #t #t #t)
        (loop (sub1 n)))))

(define (loop2 n)
  (if (zero? n)
      123456789
      (begin
        (call #t #t #t)
        (loop (sub1 n)))))

(time (loop 100000000))
(time (loop2 100000000))
```

prints

```
[samth@huor:/tmp plt] r c.rkt 
cpu time: 410 real time: 410 gc time: 0
123456789
cpu time: 408 real time: 408 gc time: 1
123456789
[samth@huor:/tmp plt] r c.rkt 
cpu time: 414 real time: 414 gc time: 0
123456789
cpu time: 412 real time: 412 gc time: 1
123456789
```