@sorawee 2022-06-24
In your `loop2`, the recursive call is to `loop`, which is the "fast" one. If you correct that mistake, `loop2` should be significantly slower than `loop`.