@samth 2022-06-24
Just adding indirect calls seems not to be enough to make things even. Here's a version of the program where the "check" version has a lot of indirect calls:

```
(module sub racket/base
  (require racket/contract/base)
  (define (call a)
    (void))

  (struct m (v))
  
  (define x #f)
  (define f #f)

  (define bbb (= 17 (random 15)))

  (set! f (lambda (n b) (if bbb
                            (error "~a ~a" n b)
                            (car (m-v n))
                            )))

  (set! x (m (list boolean?)))

  (define (call/check a b)
    (unless ((f x b) a)
      (error 'bad))
    (void))

  (provide/contract [call (-> boolean? any)])
  (provide call/check))
```

In this version, `call/check` is still about twice as fast as `call`.