@samth 2022-06-24
Oh, and the 2x ratio is similar with 3 arguments as with 1.