@rfindler 2022-06-24
Well, I eliminated one indirect call (diff below) and I see an improvement from 1974 msec to 746 msec on my machine for this loop:

```
(define (loop0 n)
  (if (zero? n)
      123456789
      (begin
        (call #t #t #t #t #t #t #t #t #t #t #t #t)
        (loop0 (sub1 n)))))
```

where I changed various things to have 12 arguments instead of 3.

Here's the diff. It may make sense to do something like this for a set of popular primitive functions, I'm not sure. Seems kind of annoying tho :)

```
diff --git a/racket/collects/racket/contract/private/guts.rkt b/racket/collects/racket/contract/private/guts.rkt
index 6b8db0acbc..b605bd12f0 100644
--- a/racket/collects/racket/contract/private/guts.rkt
+++ b/racket/collects/racket/contract/private/guts.rkt
@@ -751,6 +751,23 @@
        (procedure-closure-contents-eq? (predicate-contract-pred this)
                                        (predicate-contract-pred that))))
 
+
+(define (predicate-contract-list-contract? ctc)
+  (or (equal? (predicate-contract-pred ctc) null?)
+      (equal? (predicate-contract-pred ctc) empty?)))
+
+(define (predicate-contract-do-generate ctc)
+  (define generate (predicate-contract-generate ctc))
+  (cond
+    [generate generate]
+    [else
+     (define built-in-generator
+       (find-generate (predicate-contract-pred ctc)
+                      (predicate-contract-name ctc)))
+     (λ (fuel)
+       (and built-in-generator
+            (λ () (built-in-generator fuel))))]))
+
 ;; sane? : boolean -- indicates if we know that the predicate is well behaved
 ;; (for now, basically amounts to trusting primitive procedures)
 (define-struct predicate-contract (name pred generate sane?)
@@ -772,19 +789,28 @@
           (if (p? v)
               v
               (raise-predicate-blame-error-failure blame v neg-party name))))))
-   #:generate (λ (ctc)
-                 (let ([generate (predicate-contract-generate ctc)])
-                   (cond
-                     [generate generate]
-                     [else
-                      (define built-in-generator
-                        (find-generate (predicate-contract-pred ctc)
-                                       (predicate-contract-name ctc)))
-                      (λ (fuel)
-                        (and built-in-generator
-                             (λ () (built-in-generator fuel))))])))
-   #:list-contract? (λ (ctc) (or (equal? (predicate-contract-pred ctc) null?)
-                                 (equal? (predicate-contract-pred ctc) empty?)))))
+   #:generate predicate-contract-do-generate
+   #:list-contract? predicate-contract-list-contract?))
+
+(struct boolean?-predicate-contract predicate-contract ()
+  #:property prop:custom-write contract-custom-write-property-proc
+  #:property prop:flat-contract
+  (build-flat-contract-property
+   #:trusted trust-me
+   #:stronger predicate-contract-equivalent
+   #:equivalent predicate-contract-equivalent
+   #:name (λ (ctc) (predicate-contract-name ctc))
+   #:first-order (λ (ctc) (predicate-contract-pred ctc))
+   #:late-neg-projection
+   (λ (ctc)
+     (define name (predicate-contract-name ctc))
+     (λ (blame)
+       (λ (v neg-party)
+         (if (boolean? v)
+             v
+             (raise-predicate-blame-error-failure blame v neg-party name)))))
+   #:generate predicate-contract-do-generate
+   #:list-contract? predicate-contract-list-contract?))
 
 (define (raise-predicate-blame-error-failure blame v neg-party predicate-name)
   (raise-blame-error blame v #:missing-party neg-party
@@ -797,7 +823,7 @@
 (define (build-flat-contract name pred [generate #f])
   (make-predicate-contract name pred generate #f))
 
-(define boolean?/c (make-predicate-contract 'boolean? boolean? #f #t))
+(define boolean?/c (boolean?-predicate-contract 'boolean? boolean? #f #t))
 
 (define (contract-name ctc)
   (contract-struct-name
```